CRUD DE PROVEEDOR COMPLETO, FALTARIA DE PRODUCTO EN PROCESO Y TAMBIEN SE AGREGO LA MEJORA DE USUARIOS Y SE IMPLEMENTO EL INVENTARIO CADA UNO CON SUS REPORTES

+++ src/main/java/com/clinica/sistema/inventario/repository/ProductoRepositorio.java
@@ -3,6 +3,6 @@ package com.clinica.sistema.inventario.repository;
 import com.clinica.sistema.inventario.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepositorio extends JpaRepository<Producto, Integer> {
+public interface ProductoRepositorio extends JpaRepository<Producto, Long> {
 
 }
